# WIP: Windows: Use a faster comparison when enumeration filter strings don't have wildcards