@wilbaker 2018-10-02
Closing PR until the second commit is fixed up